requirements.txt
python-dotenv
langchain
pinecone-client
langchain-core
langchain-community
PyMuPDF
langchain-pinecone
sentence-transformers
langchain-huggingface
langchain-text-splitters
ctransformers

-e .